+++ src/main/java/org/quiltmc/loader/impl/transformer/RequiresStrippingVisitor.java
@@ -26,7 +26,7 @@ import org.objectweb.asm.MethodVisitor;
 import org.objectweb.asm.Type;
 import org.objectweb.asm.TypePath;
 import org.objectweb.asm.TypeReference;
-import org.quiltmc.loader.api.minecraft.Requires;
+import org.quiltmc.loader.api.Requires;
 import org.quiltmc.loader.api.plugin.solver.ModLoadOption;
 import org.quiltmc.loader.impl.util.QuiltLoaderInternal;
 import org.quiltmc.loader.impl.util.QuiltLoaderInternalType;
